requirements.txt
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
scikit-learn==1.4.2
wfdb==4.1.2
ultralytics
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
-f https://download.pytorch.org/whl/torch_stable.html
